trim down relay
