+++ relay/Cargo.toml
@@ -4,9 +4,13 @@ version = "0.1.0"
 edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[build-dependencies]
+ethers-solc = "1.0.2"
+
 [dependencies]
 async-trait = "0.1.58"
-bonsai-proxy-contract = { path = "src/proxy-contract" }
+bonsai-sdk-alpha = { package = "bonsai-sdk", git = "https://github.com/risc0/risc0", branch = "release-0.15" }
+bonsai-starter-methods = { path = "../methods" }
 clap = { version = "4.3", features = ["derive", "env"] }
 displaydoc = "0.2"
 ethers = { version = "=2.0.2", features = ["rustls", "ws"] }
@@ -35,4 +39,4 @@ uuid = { version = "1.3.1", features = ["v4", "serde"] }
 wiremock = "0.5"
 
 [features]
-insecure_skip_seal = []
+solidity_compile = []
