--- relay/src/downloader/bonsai/mod.rs
@@ -1,2 +1,0 @@
-pub mod proxy_callback_proof_processor;
-pub mod proxy_callback_proof_request_stream;
