--- relay/src/downloader/common/mod.rs
@@ -1,2 +1,0 @@
-pub mod event_processor;
-pub mod event_stream;
