--- relay/src/downloader/mod.rs
@@ -1,2 +1,0 @@
-pub mod bonsai;
-pub mod common;
