+++ relay/src/lib.rs
@@ -1,25 +1,26 @@
-pub mod downloader;
-pub mod storage;
-pub mod uploader;
+// Copyright 2023 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 use std::sync::Arc;
 
-use bonsai_common_log::setup_heartbeat_log;
-use downloader::bonsai::{
-    proxy_callback_proof_processor::ProxyCallbackProofRequestProcessor,
-    proxy_callback_proof_request_stream::ProxyCallbackProofRequestStream,
-};
+use bonsai_starter_methods::{resolve_guest_entry, resolve_image_output, GUEST_LIST};
 use ethers::{core::types::Address, prelude::*};
-use storage::in_memory::InMemoryStorage;
-use tokio::sync::Notify;
-use uploader::bonsai::{
-    completed_proofs::manager::BonsaiCompleteProofManager,
-    pending_proofs::manager::BonsaiPendingProofManager,
-};
+use tracing::info;
+
+abigen!(ProxyContract, "artifacts/proxy.sol/Proxy.json");
 
 pub struct Config {
-    pub bonsai_url: String,
-    pub bonsai_api_key: String,
     pub proxy_address: Address,
     pub log_status_interval: u64,
 }
@@ -29,67 +30,48 @@ where
     <M as ethers::providers::Middleware>::Provider: PubsubClient,
     <<M as ethers::providers::Middleware>::Provider as ethers::providers::PubsubClient>::NotificationStream: Sync,
 {
-    let bonsai_client = Client::new(config.bonsai_url, config.bonsai_api_key)
-        .expect("Failed to create Bonsai client.")
-        .clone();
-    let storage = InMemoryStorage::new();
+    let event_name = "CallbackRequest(address,bytes32,bytes,address,bytes4,uint64)";
+    let filter = ethers::types::Filter::new()
+        .address(config.proxy_address)
+        .event(event_name);
+    let mut proxy_stream = ethers_client
+        .subscribe_logs(&filter)
+        .await
+        .unwrap()
+        .map(|log| {
+            ethers::contract::parse_log::<CallbackRequestFilter>(log)
+                .expect("must be a callback proof request log")
+        });
 
-    // Setup Downloader
-    let new_pending_proof_request_notifier = Arc::new(Notify::new());
-    let proxy_callback_proof_request_processor = ProxyCallbackProofRequestProcessor::new(
-        bonsai_client.clone(),
-        storage.clone(),
-        Some(new_pending_proof_request_notifier.clone()),
-    );
+    let proxy: ProxyContract<M> = ProxyContract::new(config.proxy_address, ethers_client.clone());
+    while let Some(event) = proxy_stream.next().await {
+        // Search list for requested binary name
+        let image_id = hex::encode(event.image_id);
+        let guest_entry =
+            resolve_guest_entry(GUEST_LIST, &image_id).expect("Failed to resolve guest entry");
 
-    let downloader = ProxyCallbackProofRequestStream::new(
-        ethers_client.clone(),
-        config.proxy_address.clone(),
-        proxy_callback_proof_request_processor,
-    );
+        // Execute or return image id
+        let input = hex::encode(event.input);
+        let output_bytes =
+            resolve_image_output(&input, guest_entry).expect("Failed to compute journal output");
 
-    // Setup Uploader
-    let new_complete_proof_notifier = Arc::new(Notify::new());
+        // Broadcast callback transaction
+        let proof_batch = vec![Callback {
+            callback_contract: event.callback_contract,
+            journal_inclusion_proof: vec![],
+            payload: output_bytes.into(),
+            gas_limit: event.gas_limit,
+        }];
 
-    let uploader_pending_proof_manager = BonsaiPendingProofManager::new(
-        bonsai_client.clone(),
-        storage.clone(),
-        new_pending_proof_request_notifier.clone(),
-        new_complete_proof_notifier.clone(),
-    );
+        info!("sending batch");
+        let contract_call = proxy.invoke_callbacks(proof_batch);
+        let pending_tx = contract_call
+            .send()
+            .await
+            .expect("failed to send callback transaction");
 
-    let send_batch_notifier = Arc::new(Notify::new());
-    let max_batch_size: usize = 3;
-    let send_batch_interval = tokio::time::interval(tokio::time::Duration::from_millis(1000));
-
-    let uploader_complete_proof_manager = BonsaiCompleteProofManager::new(
-        bonsai_client.clone(),
-        storage.clone(),
-        new_complete_proof_notifier.clone(),
-        send_batch_notifier.clone(),
-        max_batch_size,
-        config.proxy_address.clone(),
-        ethers_client.clone(),
-        send_batch_interval,
-    );
-
-    // Start everything
-    let downloader_handle = tokio::spawn(downloader.run());
-    let uploader_pending_proof_manager_handle = tokio::spawn(uploader_pending_proof_manager.run());
-    let uploader_complete_proof_manager_handle =
-        tokio::spawn(uploader_complete_proof_manager.run());
-
-    setup_heartbeat_log(config.log_status_interval);
-
-    tokio::select! {
-        err = downloader_handle => {
-            panic!("{}", format!("downloader exited: {:?}", err))
-        }
-        err = uploader_pending_proof_manager_handle => {
-            panic!("{}", format!("pending proof manager exited: {:?}", err))
-        }
-        err = uploader_complete_proof_manager_handle => {
-            panic!("{}", format!("complete proof manager exited: {:?}", err))
-        }
+        pending_tx
+            .await
+            .expect("Failed to confirm callback transaction");
     }
 }
