--- relay/src/proxy-contract/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "bonsai-proxy-contract"
-version = "0.1.0"
-edition = "2021"
-
-[build-dependencies]
-ethers-solc = "1.0.2"
-
-[dependencies]
-ethers = { version = "=2.0.2", features = ["rustls", "ws"] }
-hex = "0.4.3"
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-tokio = { version = "1", features = ["macros", "rt"] }
-tracing = "0.1.35"
-
-[dev-dependencies]
-risc0-zkvm = { git = "https://github.com/risc0/risc0.git", branch = "release-0.15", default-features = false }
-sha2 = "0.10"
-tempfile = "3.3.0"
-
-[features]
-default = []
-solidity_compile = []
