--- relay/src/proxy-contract/artifacts/ibpn.sol/IBPN.json
@@ -1,240 +1,0 @@
-{
-  "abi": [
-    {
-      "inputs": [
-        {
-          "internalType": "bytes32",
-          "name": "image_id",
-          "type": "bytes32"
-        },
-        {
-          "internalType": "bytes32",
-          "name": "journal_hash",
-          "type": "bytes32"
-        },
-        {
-          "internalType": "bytes32[]",
-          "name": "proof",
-          "type": "bytes32[]"
-        }
-      ],
-      "name": "verify",
-      "outputs": [
-        {
-          "internalType": "bool",
-          "name": "",
-          "type": "bool"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    }
-  ],
-  "bytecode": {
-    "object": "0x",
-    "sourceMap": "",
-    "linkReferences": {}
-  },
-  "deployedBytecode": {
-    "object": "0x",
-    "sourceMap": "",
-    "linkReferences": {}
-  },
-  "methodIdentifiers": {
-    "verify(bytes32,bytes32,bytes32[])": "3423e548"
-  },
-  "ast": {
-    "absolutePath": "contracts/ibpn.sol",
-    "id": 156,
-    "exportedSymbols": {
-      "IBPN": [
-        155
-      ]
-    },
-    "nodeType": "SourceUnit",
-    "src": "69:161:2",
-    "nodes": [
-      {
-        "id": 142,
-        "nodeType": "PragmaDirective",
-        "src": "69:24:2",
-        "nodes": [],
-        "literals": [
-          "solidity",
-          "^",
-          "0.8",
-          ".17"
-        ]
-      },
-      {
-        "id": 155,
-        "nodeType": "ContractDefinition",
-        "src": "95:134:2",
-        "nodes": [
-          {
-            "id": 154,
-            "nodeType": "FunctionDefinition",
-            "src": "116:111:2",
-            "nodes": [],
-            "functionSelector": "3423e548",
-            "implemented": false,
-            "kind": "function",
-            "modifiers": [],
-            "name": "verify",
-            "nameLocation": "125:6:2",
-            "parameters": {
-              "id": 150,
-              "nodeType": "ParameterList",
-              "parameters": [
-                {
-                  "constant": false,
-                  "id": 144,
-                  "mutability": "mutable",
-                  "name": "image_id",
-                  "nameLocation": "140:8:2",
-                  "nodeType": "VariableDeclaration",
-                  "scope": 154,
-                  "src": "132:16:2",
-                  "stateVariable": false,
-                  "storageLocation": "default",
-                  "typeDescriptions": {
-                    "typeIdentifier": "t_bytes32",
-                    "typeString": "bytes32"
-                  },
-                  "typeName": {
-                    "id": 143,
-                    "name": "bytes32",
-                    "nodeType": "ElementaryTypeName",
-                    "src": "132:7:2",
-                    "typeDescriptions": {
-                      "typeIdentifier": "t_bytes32",
-                      "typeString": "bytes32"
-                    }
-                  },
-                  "visibility": "internal"
-                },
-                {
-                  "constant": false,
-                  "id": 146,
-                  "mutability": "mutable",
-                  "name": "journal_hash",
-                  "nameLocation": "158:12:2",
-                  "nodeType": "VariableDeclaration",
-                  "scope": 154,
-                  "src": "150:20:2",
-                  "stateVariable": false,
-                  "storageLocation": "default",
-                  "typeDescriptions": {
-                    "typeIdentifier": "t_bytes32",
-                    "typeString": "bytes32"
-                  },
-                  "typeName": {
-                    "id": 145,
-                    "name": "bytes32",
-                    "nodeType": "ElementaryTypeName",
-                    "src": "150:7:2",
-                    "typeDescriptions": {
-                      "typeIdentifier": "t_bytes32",
-                      "typeString": "bytes32"
-                    }
-                  },
-                  "visibility": "internal"
-                },
-                {
-                  "constant": false,
-                  "id": 149,
-                  "mutability": "mutable",
-                  "name": "proof",
-                  "nameLocation": "191:5:2",
-                  "nodeType": "VariableDeclaration",
-                  "scope": 154,
-                  "src": "172:24:2",
-                  "stateVariable": false,
-                  "storageLocation": "calldata",
-                  "typeDescriptions": {
-                    "typeIdentifier": "t_array$_t_bytes32_$dyn_calldata_ptr",
-                    "typeString": "bytes32[]"
-                  },
-                  "typeName": {
-                    "baseType": {
-                      "id": 147,
-                      "name": "bytes32",
-                      "nodeType": "ElementaryTypeName",
-                      "src": "172:7:2",
-                      "typeDescriptions": {
-                        "typeIdentifier": "t_bytes32",
-                        "typeString": "bytes32"
-                      }
-                    },
-                    "id": 148,
-                    "nodeType": "ArrayTypeName",
-                    "src": "172:9:2",
-                    "typeDescriptions": {
-                      "typeIdentifier": "t_array$_t_bytes32_$dyn_storage_ptr",
-                      "typeString": "bytes32[]"
-                    }
-                  },
-                  "visibility": "internal"
-                }
-              ],
-              "src": "131:66:2"
-            },
-            "returnParameters": {
-              "id": 153,
-              "nodeType": "ParameterList",
-              "parameters": [
-                {
-                  "constant": false,
-                  "id": 152,
-                  "mutability": "mutable",
-                  "name": "",
-                  "nameLocation": "-1:-1:-1",
-                  "nodeType": "VariableDeclaration",
-                  "scope": 154,
-                  "src": "221:4:2",
-                  "stateVariable": false,
-                  "storageLocation": "default",
-                  "typeDescriptions": {
-                    "typeIdentifier": "t_bool",
-                    "typeString": "bool"
-                  },
-                  "typeName": {
-                    "id": 151,
-                    "name": "bool",
-                    "nodeType": "ElementaryTypeName",
-                    "src": "221:4:2",
-                    "typeDescriptions": {
-                      "typeIdentifier": "t_bool",
-                      "typeString": "bool"
-                    }
-                  },
-                  "visibility": "internal"
-                }
-              ],
-              "src": "220:6:2"
-            },
-            "scope": 155,
-            "stateMutability": "view",
-            "virtual": false,
-            "visibility": "external"
-          }
-        ],
-        "abstract": false,
-        "baseContracts": [],
-        "canonicalName": "IBPN",
-        "contractDependencies": [],
-        "contractKind": "interface",
-        "fullyImplemented": false,
-        "linearizedBaseContracts": [
-          155
-        ],
-        "name": "IBPN",
-        "nameLocation": "105:4:2",
-        "scope": 156,
-        "usedErrors": []
-      }
-    ],
-    "license": "UNLICENSED"
-  },
-  "id": 2
-}
