--- relay/src/uploader/bonsai/completed_proofs/mod.rs
@@ -1,3 +1,0 @@
-mod complete_proof;
-mod error;
-pub mod manager;
