--- relay/src/uploader/bonsai/mod.rs
@@ -1,2 +1,0 @@
-pub mod completed_proofs;
-pub mod pending_proofs;
