--- relay/src/uploader/bonsai/pending_proofs/mod.rs
@@ -1,2 +1,0 @@
-pub mod manager;
-pub mod pending_proof_request_future;
